io_uring: let the connect completion carry the full user_data
Motivation:

Widening the write user_data to a long left four completions narrowing it back
to a short at the call site. connectComplete(...) was one of them even though it
never reads that parameter, which is an inconsistency with
writeComplete(byte, int, int, long) for no gain. The other three narrow without
checking, so a later change that hands one of those namespaces a wider id would
drop its high bits without a sound.

Modification:

Give connectComplete(...) a long, which removes the cast at all three of its
call sites. Route the remaining three -- read, poll and cancel -- through
narrowUserData(...), which asserts the value survives the round-trip first.
Those completions all submit ids from short-parameter factories today, so the
assert does not fire; it is there to catch a future change that stops being
true.

Result:

One narrowing point instead of four, and it fails loudly rather than quietly.

+++ transport-classes-io_uring/src/main/java/io/netty/channel/uring/IoUringDatagramChannel.java
@@ -554,7 +554,8 @@ public final class IoUringDatagramChannel extends AbstractIoUringChannel impleme
                 sendmsgHdrs.clear();
                 ChannelOutboundBuffer outboundBuffer = outboundBuffer();
                 if (outboundBuffer == null) {
-                    // The completion may arrive after close() already dropped the outbound buffer.
+                    // The completion may arrive after close() or shutdownOutput() already dropped the
+                    // outbound buffer.
                     return true;
                 }
                 for (int i = 0; i < numWritten; i++) {
@@ -584,7 +585,7 @@ public final class IoUringDatagramChannel extends AbstractIoUringChannel impleme
         }
 
         @Override
-        void connectComplete(byte op, int res, int flags, short data) {
+        void connectComplete(byte op, int res, int flags, long data) {
             if (res >= 0) {
                 connected = true;
             }
